mu-plugins: Make global terms and permalink structure WordCamp-only
